Check epsg:4326 bounds in layer creation (#317)
This adds a check on layer creation that the computed centroid of the
input data is within the bounds of wgs84. This doesn't check the
computed _bounding box_ because that could wrap over the international
date line and have values less than -180.

This is improved over the previous approach because previously we
checked the _final average_ of the centroid against wgs84, whereas now
we check each layer individually. This is safer in case _most_ layers
are wgs84 but one is not. Also, it errors at layer creation, which is a
much better time to error.

In the future, we could allow web-mercator coordinates as well (see
https://deck.gl/docs/developer-guide/coordinate-systems) and this would
have to be refactored to allow the larger range of coordinates.