Fix inferring number of rows per chunk (#327)
We use the `_rows_per_chunk` attribute to determine the number of rows
that are included in each Parquet chunk sent to the frontend. This is a
_layer-level_ construct because we need to ensure the main table and all
accessors have exactly the same chunking, because each chunk is rendered
independently as a separate deck.gl layer

We previously had issues where the number of rows per chunk was either
not the same across all data objects within a layer, or alternatively a
few cases (as with the ArcLayer) where we were accidentally initializing
the number of rows per chunk to be `0`, which let to an infinite loop in
`table.to_batches(max_chunksize=0)`.