Multi-dimensional GPU-based data filtering (#335)
Filter on up to 4 numeric attributes at the same time! This is working
locally, but I need to figure out a clean way to connect it to the
filter_value, since that's a list of lists.

The best solution (to enable jslink) might be to create our own custom
widget that can display 1-4 float sliders and stores its value as a list
of lists?