Support `pyarrow.Table` with `geoarrow.pyarrow` extension types as geometry columns (#218)
Work-in-progress, and haven't yet tested it.

The idea is that when encountering a `geoarrow.pyarrow` extension type,
to convert this to its raw storage field with metadata (i.e. how the
extension type looks like in IPC or when the python wrapper class is not
registered). Short term, this is the easiest approach, because a lot of
the existing code relies on checking eg the "ARROW:extension:name" in
the field metadata.

---------

Co-authored-by: Kyle Barron <kyle@developmentseed.org>