Sync view state between JS and Python (#448)
Closes https://github.com/developmentseed/lonboard/issues/112



In https://github.com/developmentseed/lonboard/pull/444 @batpad
mentioned

> It does seem to make the map a bit slow for me, and we should
investigate that, but in principle, this for me works as expected

While playing around with that PR, I also saw some view state lagging at
low zooms with lots of data rendered. In this screencast I'm smoothly
zooming in, but the rendering is very jagged, and sometimes zooms back
out slightly. **This screencast is of the previous PR**


https://github.com/developmentseed/lonboard/assets/15164633/bbe0a168-b76e-4563-bc30-828d648210db

My hypothesis of what's happening is that when rendering lots of data
there's a slight lag for deck to call the `onViewStateChange` handler.
And so deck will occasionally call the handler with "old" data and
that's where you sometimes get the "snap back" in the scrolling.

I think this is a really bad UX, and it seems not to happen when deck
manages the view state internally. So I think potentially the best UX
and performance solution is to decouple the view state from the model,
while still syncing from deck to Python in the debounced model handler.

This PR tries to implement this.

---------

Co-authored-by: Sanjay Bhangar <sanjay@developmentseed.org>