Fix tests with duckdb 0.10.3 (#552)
~~Need to debug this~~

Turned out this is an improvement/fix on the duckdb side to error before
doing a replacement scan across connections. Previously we had tests
that called back into the default connection, but those now fail unless
the user passes the `con` argument directly.

As long as users pass in the `con`, this just works already.