Store trait sequences as tuples, not lists (#620)
Ref
https://github.com/developmentseed/lonboard/issues/592#issuecomment-2316173005

This should be easier to follow for users. Mutating a list trait doesn't
propagate any updates to the frontend. Rather, you must set a new
object. Storing tuple traits instead of list traits means that it's
impossible for the user to accidentally not sync those updates to the
frontend.