Clicked point to Lonboard map (#671)
This PR adds a new property to the Lonboard map that holds a tuple for
the coordinate where the user last clicked on the map. It is not
implemented as an .on_click(), as was
[discussed](https://github.com/developmentseed/lonboard/discussions/633),
but I think it accomplishes the same functionality (and may actually
have more utility this way). hopefully this is an acceptable pattern,
but if not, let me know and I'll see what I can do.