Support Py3.12 in tests by constraining pyproj in lockfile to versions with wheels (#745)
This PR adds environment markers to the `uv` tooling configuration
[constraint-dependencies](https://docs.astral.sh/uv/reference/settings/#constraint-dependencies)
for `pyproj`, allowing us to have a lockfile with `pyproj` wheels for
Python versions we support without restricting versions for downstream
consumers.

I believe the path to supporting 3.13 in tests will require updating
`fiona<1.10` to something like `fiona>1,<=1.10`. It also looks safe to
bump DuckDB in our lockfile but we will have to invalidate the CI cache
so it doesn't point to extensions compiled for `1.1.3`