Scope tailwind preflight styles (#740)
Anywidget loads CSS from the _css attribute to the global scope. A
widget most take care not to conflict with any other CSS selectors. The
stylesheet from the tailwind preflight, loaded with `@tailwind base`,
will conflict with stylesheets from Sphinx themes (#698) or the VScode
interactive display (#696).

To scope the preflight stylesheet to descendants of a wrapper should fix
that.

---------

Co-authored-by: Kyle Barron <kyle@developmentseed.org>