feat: Support passing `ChunkedArray`/`ArrayReader` into `viz` (#824)
### Change list

Currently, passing a `ChunkedArray`/`ArrayReader` into `viz` fails
because any `ArrowStreamExportable` it tries to import as a table.

This adds a helper for checking the field of an input stream. If the
field is struct but **not** a GeoArrow extension type, then it assumes
the input is a `Table`. We need to check for the extension type because
that ensures that we can render a struct-encoded point column or a
`geoarrow.box` column.

Also:

- Update `test_viz.py` for latest geoarrow-rust-core version 0.4.
- Rename `geoarrow` test folder so that when starting a Python kernel in
the above directory it doesn't search in the test files for the
`geoarrow` python library.
- Update some other dev deps

https://github.com/developmentseed/lonboard/issues/823 will be a follow
up