feat: Use `100%` as default CSS height (#835)
This PR will replace the existing private `_height` property of a
Lonboard map with a new public `height` property. The new property will
accept an integer (assumed to be the number of pixels) for the map, or a
string that contains a CSS height (for example, '100%' or '24rem')

When lonboard maps are exported to an HTML file using the `.to_html()`
method, the map in the resulting HTML file will always be set to 100%
height so the map will fill the entire browser window.

I believe this will resolve #781 since the new property doesn't have the
preceding underscore. as well as #729

---------

Co-authored-by: Kyle Barron <kyle@developmentseed.org>