feat: Support two render modes: Standard/deck.gl-first and MapboxOverlay (#921)
As described in the [upstream deck.gl
docs](https://deck.gl/docs/developer-guide/base-maps/using-with-maplibre),
there are three ways to support using deck.gl with Maplibre:
_interleaved_, _overlaid_, and _reverse-controlled_. The first two are
supported via `MapboxOverlay` with a prop `interleaved: true|false`,
while the latter is implemented by having Maplibre be a child of the
deck.gl Map.

There are worthwhile reasons to support all of these modes. We need to
use interleaved or overlaid to support globe view, while
reverse-controlled better supports multiple deck.gl views.

This PR refactors the map component in `index.tsx` into two separate
React components: a deck.gl-first renderer (i.e. "reverse-controlled")
and a MapboxOverlay renderer.

The idea is that this will pair with #908 to give users more control
over various ways of rendering maps.

This is backwards-compatible because we default to reverse-controlled,
the existing default.

Closes https://github.com/developmentseed/lonboard/pull/890, closes
https://github.com/developmentseed/lonboard/pull/437, for
https://github.com/developmentseed/lonboard/issues/886, for
https://github.com/developmentseed/lonboard/issues/718