refactor: Refactor model initialization logic (#923)
In ipywidgets, a "model" is the core of reactivity. In particular, a
Lonboard `Map` is reactive, but every underlying layer is _also_
reactive, so that when a user changes any property on a layer, it
automatically gets updated on the map.

By extension, our
[`Extension`](https://developmentseed.org/lonboard/latest/api/layer-extensions/)
classes are also themselves models. These are separate Python classes
that get serialized as their own reactive models on the JS side.

Soon we'll have more types of widget models that we want to be reactive,
including views and basemap styles (#908).

Right now, model initialization is slightly hard-coded to handle layer
initialization. This PR refactors the model initialization to be fully
generic.