feat: Create richer `Basemap` class and deprecate `basemap_style` arg (#935)
### Change list

- Add `basemap` parameter to `Map`. This now manages a class with more
information than just the basemap style. In particular, this holds the
render mode `"interleaved"`, `"overlaid"`, or `"reverse-controlled"`
that was implemented on the JS side in
https://github.com/developmentseed/lonboard/pull/921
- Deprecate `basemap_style` parameter to `Map`. This is superseded by
the `style` parameter to `MaplibreBasemap`
- Rename `CartoBasemap` to `CartoStyle`
- Add `before_id` parameter to the core `Layer`. When passed and the
basemap render mode is `"interleaved"`,
- Pass `interleaved` prop down to MapboxOverlay
- Remove temporary `render_mode` from #921

**Rendering layer interleaved in maplibre layer stack:**

<img width="817" height="505" alt="image"
src="https://github.com/user-attachments/assets/c63148a9-daa0-4dd4-bb65-88cc13805060"
/>

TODO:

- [ ] Validation for `beforeId`. If `beforeId` is passed to a layer, try
to resolve the basemap style URL to a dict, and then validate that the
string value of `beforeId` exists in that layer stack. Keep a cache of
fetched styles so that you're not fetching them repeatedly.
- [ ] Warn if `beforeId` is set when the basemap style is not
`"interleaved"`?
- [ ] (next PR): integration with
https://github.com/geopandas/xyzservices (see #494)
- [ ] (possibly future PR): examples of rendering different basemap
modes

For https://github.com/developmentseed/lonboard/issues/494

---------

Co-authored-by: Vitor George <vitor.george@gmail.com>