feat: Validate that GlobeView is only used with interleaved basemap mode (#1012)
After discussion with @felixpalmer , we established that the only
deck.gl rendering mode that really supports Globe view is
interleaved-maplibre (or no basemap at all).

This adds validators to the `Map` class to enforce that you can't add a
GlobeView when the basemap is overlaid or reverse-controlled, and you
can't set such a basemap when the view is a globe view.

ref https://github.com/developmentseed/lonboard/issues/1009